requirements.txt
supabase
python-dotenv
flask
